airflow/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cloud-sql-python-connector==0.8.0
cryptography==37.0.4
frozenlist==1.3.1
google-auth==2.10.0
google-api-core==2.8.0
google-auth-oauthlib==0.5.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.1
idna==3.3
multidict==6.0.2
pandas==1.3.5
pg8000==1.29.1
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.20.0
requests==2.28.1
rsa==4.9
scramp==1.4.1
six==1.16.0
urllib3==1.26.11
yarl==1.8.1
